Configure initial CLI
This sets up a basic CLI which can be run with `uv run
govuk_chat_evaluation` after running `uv pip install -e .` to make the
script available.

By virtue of the __main__.py file in govuk_chat_evaluation this can also
be run directly as a python module with `uv run -m govuk_chat_evaluation`
which is useful for if there are any problems with the `pip install
command`.

Choice of uv as the python runner is based on the GDS Way [1]

[1]: https://gds-way.digital.cabinet-office.gov.uk/manuals/programming-languages/python/python.html#environments