Add basic CI configuration
This adds a few CI tasks for code formatting: `ruff check` for linting,
`ruff format` for compliance with auto-formatting and `pyright` for
following typing consistency.

The choice of `ruff` is based on the GDS Way [1] but I've not configured
it (kinda hoping we don't need to) whereas pyright was chosen for typing
just because that was what my text editor was already configured for.

Pyright can be a bit annoying as Python seems to be still early in it's
journey towards types and it seems to be easy to hit complex situations
that are hard to resolve - we'll see how it pans out.

It also uses direct hashes for the GitHub Actions as per the GDS Way
[2].

[1]: https://gds-way.digital.cabinet-office.gov.uk/manuals/programming-languages/python/python.html#linting
[2]: https://gds-way.digital.cabinet-office.gov.uk/standards/source-code/use-github.html#using-github-actions-and-workflows