Config module for command line arguments
This creates a class that can be used to represent the config of an
evaluation. This class has helpers to interface with click so that the
attributes of the config model can be applied as arguments to the
evaluation.

It's pulled out of a wider piece of work to create a more granular
commit so full purpose isn't apparent at this stage.

This also introduces the first unit tests for this project and applies
the CI configuration.