Add dataset_generation module
This module is provided as a means to asynchronously generate data, with
the generate_dataset being an async function that'll run a function for
every ground truth in a list; and a function to asynchronously run a
rake task on GOV.UK Chat as the expected means to be used to generate
data.

There's a bunch of scope for improvement on this code both the module
and the tests.

The rake task is the simpler of the code, it is hard coded to a
particular path for GOV.UK Chat and a particular set of commands to run
the task. In practice we may need this to be configurable, particularly
if we were to consider running the rake task directly on machines in the
K8s cluster. This function isn't too complex to understand though
(mercilessly).

The generate_dataset is more complicated and I struggled with how to
test it fully (particularly cancelling incomplete jobs). This uses a
semaphore as a means to set a limit on concurrency which is hard coded at
10 and may need to be configurable.