Add minimal jailbreak guardrails evaluation
This adds a subcommand to the govuk_chat_evaluation task which has the
job to perform the jailbreak guardrails evaluation.

I've set this up as a minimal task which merely loads the config file,
creates an output directory and outputs the used config to that file.

Expected usage of this would be:

uv run govuk_chat_evaluation jailbreak_guardrails

We can adjust the config file for it with:

uv run govuk_chat_evaluation jailbreak_guardrails --config_path "config.yaml"

Or to change the what value:

uv run govuk_chat_evaluation jailbreak_guardrails --what "A different value"