Add code to evaluate jailbreak guardrails
This adds some basic evaluation code for jailbreak guardrails. It serves
mostly as an example of what we can do with evaluation rather than being
what we actually need - demonstrating a model for each result, which is
wrote to a CSV, and an aggregates model for representing the
aggregations.

The output CSV for aggregations is very speculative - I think we'll
probably need something quite different than a CSV with headers of
property and value.