Add ability to generate data from GOV.UK Chat
This makes use of the functions from the govuk_chat_evaluation.generate
module to asynchronously run rake tasks on govuk-chat to establish
output.

When this output is generated it is wrote to a file, which allows
evaluating it again without a need to generate it again.

This adds a generate and provider argument to jailbreak_guardrails
Config to allow controlling generation and arguments to the generation.