Use None as default option for bool CLI options
This corrects a problem where not specifying a boolean argument forced
it to be false and thus only allowed setting it to be true with a CLI
argument.

The above sounds quite opaque so I'll demonstrate with an example:

If we have a config file for jailbreak guardrails with a `generate: true` and run:

```
uv run govuk_chat_evaluation jailbreak_guardrails
```

Then this overrides the `generate: true` to `generate: false` because of
the default value of the CLI argument.

By setting this to a default of None, it only has an affect when the
argument is specified. This is the behaviour we want for appropriately
cascading options.