Add python-dotenv for loading env vars
This is to use a .env file to load env vars that'll be used - with the
expectation that these will be typically used as credentials for LLMs.

I've added the load_dotenv to the root cli file as I believe that will
be called by any usage of the CLI interface.