Add mechanism to reuse common fields
I've put this in with an expectation that we may reuse the same fields
multiple times and that it would be better to have some sort of
namespace to reference them.

I'm not too sure if the approach I've taken here is idiomatic python
though (it was quite a challenge to work out the approach).