Mechanism to generate rag answers
This adds the code to interface with the Ruby rake task in order to
generate an answer. It is quite similar to the code of the same purpose
for jailbreak_guardrails.

It's pertinent to note that this is likely the absolute minimum of data
for usage. I expect to use this properly we'll have to consider a more
complex data structure and likely validate it (so probably need a
distinct Pydantic model).