Configure generation as part of RAG rag_answers
This code is added so that a rag_answers evaluation can also generate
data. It's implementation is based heavily on the equivalent of
jailbreak_guardrails.

I thought for testing "--no-generate", it'd be a bit more honest to have
the default set to generate so the test shows it overriding, whereas
previously it could also be showing no effect.