Add evaluation for question routing
https://trello.com/c/82aouRre/2377

This extends the evaluation to support question routing.

It works in much the same way as the existing evaluations in that it
calls the rake task from the Rails app to generate the evals, then
outputs the results in the same way. However there are a couple of
differences.

Firstly we output a confusion matrix diagram for the eval. This uses the
sklearn, seaborn and matplotlib dependencies to generate the image. It's
a visual way to quickly see how well the eval is performing.

Secondly we output a CSV of miscategorised cases, i.e. it just contains
the entries of questions which were classified incorrectly.