Refactor evaluate tests for rag_answers
This adjusts the approach to mocking to be more consistent with other
files in this repository and be a bit simpler.

The approach taken is to try focus on just mocking what is in the unit
of code we are testing and seeing where we can use existing classes and
mocks for realism.

I've adjusted a bit of what we assert on, where I've removed some of the
assertions that were perhaps testing underlying libraries (such as in
AggregatedResults) and increased testing of the unit in the test (such
as in asserting that CSV files are CSVs with headers).

I did hit an annoying issue with the Pydantic MetricData model from
Deepeval where Pyright required optional arguments (strictMode,
evaluationModel, etc) which are aliases [1] - I understand this is an
quirk of Pydantic and Pyright that can be fixed in the model but as I
didn't have access to that I used a pyright ignore.

[1]: https://github.com/confident-ai/deepeval/blob/7e329c2501b3bd341daca8a6dd658f42134cdf63/deepeval/tracing/api.py#L21