Simplify the deepeval_evaluate tests
This reduces the amount of mocking and assertions that are taking place
in the deepeval_evaluate tests to make them a bit easier to understand.
It adjusts the mock_deepeval_results mock so that input is consistent
to a run but metrics vary slightly.

This also fixes the issue where we needed to configure pytest to ignore
TestResult.