Simplify test_cli for rag_answers
This reduces the amount of set up needed to run the tests for CLI by
relying on mocks that just target the key bits that call third party
services.

I moved the mock_deepeval_evaluate method to conftest so it could be
used by both test_cli and test_deepeval_evaluate.