Refactor jailbreak guardrails for handling error responses
This changes the handling of the
evaluation:generate_jailbreak_guardrail_response rake task to reflect
that it now returns success or response_error as changed in
https://github.com/alphagov/govuk-chat/pull/200.

It is now configured so that when an response_error is returned the
value of this is logged and it isn't included in the results.

A notable change of this is that it is now somewhat possible to get the
results where none can be processed - which is a situation that
currently errors.