Rename logfile from run.log to problems.log
In discussions for
https://github.com/alphagov/govuk-chat-evaluation-prototype/pull/30 we
debated whether we'd record all log information in the log file or just
warnings/errors. We decided we'd name it run.log when we thought it'd
have all the information and didn't then rename it when we decided on
only storing warnings/errors. This renames it to be a big more
descriptive.