Configure uv as a package
We can avoid the manual step of `uv pip install -e .` by flagging to uv
that this project is a package.

This also resolves a warning that was being flagged on `uv sync` of:

```
warning: Skipping installation of entry points (`project.scripts`) because this project is not packaged; to install entry points, set `tool.uv.package = true` or define a `build-system`
```

As we had a choice of how to resolve this warning,  I went with the
simpler uv approach given the expectation is that people use `uv` to run
this project.