Refactor output guardrails based on updated rake task
The `govuk-chat` output guardrails were updated in
 https://github.com/alphagov/govuk-chat/pull/168
in order to provide the name and result of each guardrail definition
for this evaluation. This has allowed for a major refactor of the
output guardrails evaluation module.

Key changes:

- all parsing logic has been removed as this is not taken care of in
the GOV.UK Chat codebase
- dynamically calculate number of guardrails rather than use a
hard-coded constant
- output metrics now use dynamically use actual guardrail names
sourced from the rake task, rather than numbers