Retrieval: add generation class
This adds the basic class for generating the results for retrieval by
the Rails application.

It follows the standard pattern for the other metrics in that we call
the Rake task and return a list of `EvaluationResult` objects.

The Rake task takes the input question and retrieves the relevant search
results from OpenSearch. It then aggregates the `exact_path` for each
result and returns them as an array.

We then create the `EvaluationResult` from the input question, the
`expected_exact_paths` that come from the evaluation input, and the
`actual_exact_paths` returned from the Rake task. The actual evaluation
will happen in a later commit.