Retrieval: add evaluation class
This adds the core evaluation metrics for retrieval. It differs slightly
to how we evaluate the other areas (question routing, guardrails, etc).

For each `EvaluationResult` we have a few properties:

* The input question
* The `expected_exact_paths`, i.e. an array of GOV.UK paths that we
  expect to be returned from Chat when it queries OpenSearch with the
  question
* The `actual_exact_paths` that were returned

Rather than calculating precision, recall, f1 and f2 scores on the
aggregated list of `EvaluationResult` objects like we do elsewhere, we
want to calculate those values on each individual `EvaluationResult`
object.

We'll then include those metrics in the results.csv file that's
generated.

For aggregate evaluation - i.e. evaluation on the whole list of result
objects - we calculate the mean, median, max and standard deviation for
each of the metrics. It's these values we'll include in the
aggregate.csv file.

For ease of readability and testing, we round all values to 4 decimal
places.